Broke up Inheritance into separate services
