+++ src/Schema/DataObject/InheritanceUnionBuilder.php
@@ -1,0 +1,161 @@
+<?php
+
+
+namespace SilverStripe\GraphQL\Schema\DataObject;
+
+
+use SilverStripe\Core\Injector\Injectable;
+use SilverStripe\GraphQL\Schema\Exception\SchemaBuilderException;
+use SilverStripe\GraphQL\Schema\Field\ModelQuery;
+use SilverStripe\GraphQL\Schema\Schema;
+use SilverStripe\GraphQL\Schema\SchemaConfig;
+use SilverStripe\GraphQL\Schema\Type\ModelInterfaceType;
+use SilverStripe\GraphQL\Schema\Type\ModelUnionType;
+use SilverStripe\ORM\DataObject;
+use ReflectionException;
+
+/**
+ * A schema-aware services for DataObject model types that creates union types
+ * for all the members of an inheritance chain. Can also apply these unions
+ * to queries to enforce unions when return types have descendants.
+ */
+class InheritanceUnionBuilder
+{
+    use Injectable;
+
+    /**
+     * @var Schema
+     */
+    private $schema;
+
+    /**
+     * InheritanceUnionBuilder constructor.
+     * @param Schema $schema
+     */
+    public function __construct(Schema $schema)
+    {
+        $this->schema = $schema;
+    }
+
+    /**
+     * @return void
+     * @throws ReflectionException
+     * @throws SchemaBuilderException
+     */
+    public function createUnions(): void
+    {
+        $dataObjects = $this->getSchema()->getModelTypesFromClass(DataObject::class);
+        $schema = $this->getSchema();
+
+        /* @var ModelInterfaceType $interface */
+        foreach ($dataObjects as $modelType) {
+            $chain = InheritanceChain::create($modelType->getModel()->getSourceClass());
+            $name = static::unionName($modelType->getName(), $schema->getConfig());
+            $union = ModelUnionType::create($modelType, $name);
+
+            $types = array_map(function ($class) use ($schema) {
+                return $schema->getConfig()->getTypeNameForClass($class);
+            }, $chain->getInheritance());
+
+            $types[] = $modelType->getName();
+
+            $union->setTypes($types);
+            $union->setTypeResolver([AbstractTypeResolver::class, 'resolveType']);
+            $schema->addUnion($union);
+        }
+    }
+
+    /**
+     * Changes all queries to use inheritance unions where applicable
+     * @throws SchemaBuilderException
+     */
+    public function applyUnions(): void
+    {
+        $schema = $this->getSchema();
+        $queries = [];
+        foreach ($schema->getQueryType()->getFields() as $field) {
+            if ($field instanceof ModelQuery) {
+                $queries[] = $field;
+            }
+        }
+        foreach ($schema->getModels() as $model) {
+            foreach ($model->getFields() as $field) {
+                if ($field instanceof ModelQuery) {
+                    $queries[] = $field;
+                }
+            }
+        }
+        foreach ($schema->getInterfaces() as $interface) {
+            if (!$interface instanceof ModelInterfaceType) {
+                continue;
+            }
+            foreach ($interface->getFields() as $field) {
+                if ($field instanceof ModelQuery) {
+                    $queries[] = $field;
+                }
+            }
+        }
+        /* @var ModelQuery $query */
+        foreach ($queries as $query) {
+            $typeName = $query->getNamedType();
+            $modelType = $schema->getModel($typeName);
+            // Type was customised. Ignore.
+            if (!$modelType) {
+                continue;
+            }
+            if (!$modelType->getModel() instanceof DataObjectModel) {
+                continue;
+            }
+
+            $unionName = static::unionName($modelType->getName(), $schema->getConfig());
+            if ($union = $schema->getUnion($unionName)) {
+                $query->setNamedType($unionName);
+            }
+        }
+    }
+
+    /**
+     * @param string $modelName
+     * @param SchemaConfig $schemaConfig
+     * @return string
+     * @throws SchemaBuilderException
+     */
+    public static function unionName(string $modelName, SchemaConfig $schemaConfig): string
+    {
+        $callable = $schemaConfig->get(
+            'inheritance.union_formatter',
+            [static:: class, 'defaultUnionFormatter']
+        );
+        return $callable($modelName);
+    }
+
+    /**
+     * @param string $modelName
+     * @return string
+     */
+    public static function defaultUnionFormatter(string $modelName): string
+    {
+        return $modelName . 'InheritanceUnion';
+    }
+
+
+    /**
+     * @return Schema
+     */
+    public function getSchema(): Schema
+    {
+        return $this->schema;
+    }
+
+    /**
+     * @param Schema $schema
+     * @return InheritanceUnionBuilder
+     */
+    public function setSchema(Schema $schema): InheritanceUnionBuilder
+    {
+        $this->schema = $schema;
+        return $this;
+    }
+
+
+}
